checkpoint 1 finished

+++ CS321_W5D2_BlogAPI.Core/Services/PostService.cs
@@ -19,10 +19,24 @@ namespace CS321_W5D2_BlogAPI.Core.Services
 
         public Post Add(Post newPost)
         {
-            // TODO: Prevent users from adding to a blog that isn't theirs
+            // Prevent users from adding to a blog that isn't theirs
             //     Use the _userService to get the current users id.
             //     You may have to retrieve the blog in order to check user id
-            // TODO: assign the current date to DatePublished
+            // assign the current date to DatePublished
+            Blog blog = _blogRepository.Get(newPost.BlogId);
+            if (blog == null)
+            {
+                throw new ApplicationException("Cannot find the requested blog.");
+            }
+
+            var postUserID = blog.UserId;
+
+            if (_userService.CurrentUserId != postUserID)
+            {
+                throw new ApplicationException("You can only make posts to your blog.");
+            }
+
+            newPost.DatePublished = DateTime.Now;
             return _postRepository.Add(newPost);
         }
 
@@ -44,13 +58,22 @@ namespace CS321_W5D2_BlogAPI.Core.Services
         public void Remove(int id)
         {
             var post = this.Get(id);
-            // TODO: prevent user from deleting from a blog that isn't theirs
+            // prevent user from deleting from a blog that isn't theirs
+            if (post.Blog?.UserId != _userService.CurrentUserId)
+            {
+                throw new ApplicationException("You can only modify blogs that you created.");
+            }
             _postRepository.Remove(id);
         }
 
         public Post Update(Post updatedPost)
         {
-            // TODO: prevent user from updating a blog that isn't theirs
+            var postUserId = updatedPost.Blog?.UserId;
+            // prevent user from updating a blog that isn't theirs
+            if (postUserId != _userService.CurrentUserId)
+            {
+                throw new ApplicationException("You can only modify blog posts that you created.");
+            }
             return _postRepository.Update(updatedPost);
         }
